move request tests into own class
testMiddlewareIsOnlyAppliedOnce maybe needs to be tested differently
